+++ tests/input_converter/test_converter.py
@@ -10,17 +10,10 @@
 #
 # This file is part of the Antares project.
 
-import pandas as pd
 import pytest
-from antares.craft.model.area import Area
 from antares.craft.model.study import Study
-from antares.craft.model.thermal import ThermalCluster
 
 from andromede.input_converter.src.converter import AntaresStudyConverter
-from andromede.input_converter.src.data_preprocessing.thermal import (
-    Direction,
-    ThermalDataPreprocessing,
-)
 from andromede.input_converter.src.logger import Logger
 from andromede.input_converter.src.utils import transform_to_yaml
 from andromede.study.parsing import (
@@ -818,241 +811,3 @@ class TestConverter:
             expected_link_component, key=lambda x: x.id
         )
         assert links_connections == expected_link_connections
-
-    def _setup_preprocessing_thermal(
-        self, local_study_w_thermal: Study
-    ) -> ThermalDataPreprocessing:
-        """
-        Initializes test parameters and returns the instance and expected file path.
-        """
-
-        logger = Logger(__name__, local_study_w_thermal.service.config.study_path)
-        converter: AntaresStudyConverter = AntaresStudyConverter(
-            study_input=local_study_w_thermal, logger=logger
-        )
-
-        def _get_thermal_prepro_with_first_thermal(
-            converter: AntaresStudyConverter, area_id: str = "fr"
-        ) -> ThermalDataPreprocessing:
-            areas: dict[Area] = converter.study.get_areas().values()
-
-            thermal: ThermalCluster = next(
-                (
-                    thermal
-                    for area in areas
-                    for thermal in area.get_thermals().values()
-                    if thermal.area_id == area_id
-                ),
-                None,
-            )
-            return ThermalDataPreprocessing(thermal, converter.study_path)
-
-        return _get_thermal_prepro_with_first_thermal(converter)
-
-    def _validate_component(
-        self,
-        component: InputComponentParameter,
-        component_id,
-        expected_path,
-        expected_values,
-    ):
-        """
-        Executes the given processing method, validates the component, and compares the output dataframe.
-        """
-
-        expected_component = InputComponentParameter(
-            id=component_id,
-            time_dependent=True,
-            scenario_dependent=True,
-            value=str(expected_path).removesuffix(".txt"),
-        )
-        current_df = pd.read_csv(expected_path, header=None)
-        expected_df = pd.DataFrame(expected_values)
-        assert current_df.equals(expected_df)
-        assert component == expected_component
-
-    @pytest.mark.parametrize(
-        "local_study_w_thermal",
-        [
-            (
-                pd.DataFrame(
-                    [
-                        [1, 1, 1, 2],
-                        [2, 2, 2, 6],
-                        [3, 3, 3, 1],
-                    ]
-                ),  # modulation
-                pd.DataFrame(
-                    [
-                        [8],
-                        [10],
-                        [2],
-                    ]
-                ),  # series
-            ),
-        ],
-        indirect=True,
-    )
-    def test_p_min_cluster(self, local_study_w_thermal):
-        """Tests the p_min_cluster parameter processing."""
-        tdp = self._setup_preprocessing_thermal(local_study_w_thermal)
-        expected_path = (
-            local_study_w_thermal.service.config.study_path
-            / "input"
-            / "thermal"
-            / "series"
-            / "fr"
-            / "gaz"
-            / "p_min_cluster.txt"
-        )
-        expected_values = [
-            [4.0],
-            [10.0],
-            [2.0],
-        ]  # min(min_gen_modulation * unit_count * nominal_capacity, p_max_cluster)
-        component = tdp.generate_component_parameter("p_min_cluster")
-        self._validate_component(
-            component, "p_min_cluster", expected_path, expected_values
-        )
-
-    @pytest.mark.parametrize(
-        "local_study_w_thermal",
-        [
-            (
-                pd.DataFrame(
-                    [
-                        [1, 1, 1, 2],
-                        [2, 2, 2, 6],
-                        [3, 3, 3, 1],
-                    ]
-                ),  # modulation
-                pd.DataFrame(
-                    [
-                        [8],
-                        [10],
-                        [2],
-                    ]
-                ),  # series
-            ),
-        ],
-        indirect=True,
-    )
-    def test_nb_units_min(self, local_study_w_thermal: Study):
-        """Tests the nb_units_min parameter processing."""
-        tdp = self._setup_preprocessing_thermal(local_study_w_thermal)
-        expected_path = (
-            local_study_w_thermal.service.config.study_path
-            / "input"
-            / "thermal"
-            / "series"
-            / "fr"
-            / "gaz"
-            / "nb_units_min.txt"
-        )
-        expected_values = [[2.0], [5.0], [1.0]]  # ceil(p_min_cluster / p_max_unit)
-
-        tdp.generate_component_parameter("p_min_cluster")
-        component = tdp.generate_component_parameter("nb_units_min")
-
-        self._validate_component(
-            component, "nb_units_min", expected_path, expected_values
-        )
-
-    @pytest.mark.parametrize(
-        "local_study_w_thermal",
-        [
-            (
-                pd.DataFrame(
-                    [
-                        [1, 1, 1, 2],
-                        [2, 2, 2, 6],
-                        [3, 3, 3, 1],
-                    ]
-                ),  # modulation
-                pd.DataFrame(
-                    [
-                        [8],
-                        [10],
-                        [2],
-                    ]
-                ),  # series
-            ),
-        ],
-        indirect=True,
-    )
-    def test_nb_units_max(self, local_study_w_thermal: Study):
-        """Tests the nb_units_max parameter processing."""
-        tdp = self._setup_preprocessing_thermal(local_study_w_thermal)
-        expected_path = (
-            local_study_w_thermal.service.config.study_path
-            / "input"
-            / "thermal"
-            / "series"
-            / "fr"
-            / "gaz"
-            / "nb_units_max.txt"
-        )
-        expected_values = [[4.0], [5.0], [1.0]]  # ceil(p_max_cluster / p_max_unit)
-
-        tdp.generate_component_parameter("p_min_cluster")
-        component = tdp.generate_component_parameter("nb_units_max")
-
-        self._validate_component(
-            component, "nb_units_max", expected_path, expected_values
-        )
-
-    def nb_units_max_variation(
-        self,
-        local_study_w_thermal: Study,
-        direction: Direction,
-    ):
-        """
-        Tests nb_units_max_variation_forward and nb_units_max_variation_backward processing.
-        """
-
-        tdp = self._setup_preprocessing_thermal(local_study_w_thermal)
-        expected_path = (
-            local_study_w_thermal.service.config.study_path
-            / "input"
-            / "thermal"
-            / "series"
-            / "fr"
-            / "gaz"
-            / f"nb_units_max_variation_{direction.value}.txt"
-        )
-        tdp.generate_component_parameter("nb_units_max")
-
-        variation_component = tdp.generate_component_parameter(
-            f"nb_units_max_variation_{direction.value}"
-        )
-
-        current_df = pd.read_csv(variation_component.value + ".txt", header=None)
-
-        nb_units_max_output = pd.read_csv(
-            tdp.series_path / "nb_units_max.txt", header=None
-        )
-
-        assert current_df[0][0] == max(
-            0, nb_units_max_output[0][167] - nb_units_max_output[0][0]
-        )
-        assert current_df[0][3] == max(
-            0, nb_units_max_output[0][2] - nb_units_max_output[0][3]
-        )
-        assert current_df[0][168] == max(
-            0, nb_units_max_output[0][335] - nb_units_max_output[0][168]
-        )
-
-        assert variation_component.value == str(expected_path).removesuffix(".txt")
-
-    @pytest.mark.parametrize(
-        "direction, local_study_w_thermal",
-        [
-            (Direction.FORWARD, DATAFRAME_PREPRO_THERMAL_CONFIG),
-            (Direction.BACKWARD, DATAFRAME_PREPRO_THERMAL_CONFIG),
-        ],
-        indirect=["local_study_w_thermal"],
-    )
-    def test_nb_units_max_variation(
-        self, local_study_w_thermal: Study, direction: Direction
-    ):
-        self.nb_units_max_variation(local_study_w_thermal, direction)
